[INTERNAL] sap.ui.fl - delegate mediator extended by requiredLibraries
For default delegates we need to add the information about the libraries
that are required to be loaded before the addViaDelegate change gets
applied. They can be now definded during the default delegate registration.
Additionally the information in which lib the default delegates are
located is added to the delegate mediator. You can catch them via
new getter function.

Change-Id: Id681c677247642aecce3099a7a40784dc10d2d3e
Signed-off-by: Arthur Trauter <arthur.trauter@sap.com>
JIRA: UICSFLEX-7268