[INTERNAL] m.Table: prevent activeHeader rendering when InvisibleText control is used as header
When the table contains active header but there is a column with InvisibleText
control set to the its header aggregation, then this would lead to an additional
tab stop in the column header which is not expected. So the table now checks
explicitly for the above mentioned setup and prevent the active header from rendering.
BCP: 2080222487

Change-Id: I731a5b85a14460ad45f380d6b72c4a13c8f3aa9d