[FIX] Field: if same value is entered again, it is ignored
There is an optimization in oData V4. If a filter request is triggered
twiche, no new backend request is sent as the result is known. Therefore
the "dataReceived" event of the ListBinding is never triggered.
In the V4 case now the "change" event is used to determine the filter.
Somehow this is not fired in V2 cases, so there still the "dataReceived"
event is used. This needs to be investigated later.

Change-Id: I7fb62730a6431dabfe70b842ad873d81b6f5eeb6
BCP: 2080249445