[INTERNAL] sap.ui.integration.widgets.Card: ChoiceSet placeholder removed from Adaptive Cards
Since implementing the static area in UI5 WC, there is no way to 'reach' the popup of a ui5-select
web component and to overwrite the css styles. This is an issue, since we used to
hide the first item of the list in order to use it as a placeholder for the ui5-select with custom CSS.
Now the 'placeholder' item is always visible in the DOM and it is confusing for the user.

Change-Id: I7f8f82af5b656de076babdd436b2070aec764a44
JIRA: BGSOFUIRILA-2866