[INTERNAL] ResizeHandler: static isSuspended with callback
This can be used to check (also from outside parties using the
ResizeObserver) whether resize handling is suspended on a certain DOM
node.
Usually the resume() after a suspend() comes after the final sizes have
been reached. The traditional ResizeHandler enforces a size check then,
to allow adapting to the new dimensions. Direct users of the
ResizeObserver, however, will not be informed and would lack the final
adaptation.
Hence, a callback is introduced to help cleaning up immediately after
the size change has finished and resume() was called.
The callback is called once when the DOM node is resumed, which caused
the suspension. This still does not necessarily mean that there is NO
DOM node causing a suspended state anymore.
Multiple calls to isSuspended may lead to multiple invocations of the
callback when equality cannot be detected.

Change-Id: I725b76694522c7432326e0921cce8dc5afa471b3