[FIX] sap.m.TimePicker: on focus in behavior over the input is now consistent
Problem description:
When "setValue" method is called with a value string,
then internal control logic format this string according
to the display format. The mask temporary value is set
to the given by the user value, which may differ from
the display value. In this case the mask temporary value
indicates that there are still unfilled input characters,
which is not true, and puts a caret into the input field.

Solution:
The mask input temporary value is reset inside the "setValue" method.

Change-Id: I4ae908df337ebc03b3c99888dc4f7052f9fd2b21
CR-Id: 002075125900003011272020
BCP: 002075129500003307462020
(cherry picked from commit 7554fabcbc5f700ee47495163b3cdb377403206c)