[INTERNAL] m.ListBase: fix sticky scroll handling
The ListBase ensure that the focused item is fully visible when sticky
feature is enabled. For this it uses the ScrollDelegate.scrollToElement
method. The ListBase was remembering the last focused item, so for the first
time when the user foused on the item it works as expected and then when the
user does the Shift+Tab which take a focus of the List and Tabs again into
the list the page scrolls down a bit and the item is not fully visible. Hence
to fix the issue, the list does not remember the last focused item and enforces
to scroll to ensure that the item is always fully visible for the user.
BCP: 2080179380

Change-Id: Ia2f82cb9e859fe466235dcc8925a2b56a5858605