[FIX] sap.m.Wizard: Next button handler now fired before attachComplete
The next button handler now is internally fired after the complete event
without using the attachEvent method of the control.

Previously _handleNextButtonPress was exectued as attachEvent handler natively, causing
it to be always fired before any external handlers attached for the 'complete'
event.

This was making setting a new step with attachComplete event from outside
impossible, as we were checking for next step in our next button handler,
that was always getting called first as attachComplete handler.

BCP: 2080264920
Change-Id: I0ba27dd34dfda490efc88bf50c6e5d84c96e4d4b