[INTERNAL] sap.ui.integration: Introduce validators concept
This change introduces a generic validation concept to replace
the previous, editor-specific validation logic.
Inside the editor default config and the property config it is
now possible to include validator functions and configure them
for the specific usecase.
Validators can be registered by the editor itself or via the config
and will run for each new value which is set.

Change-Id: Ib5db0905bb17854db2f9b7130ad03d11548ab47c
JIRA: UICSFLEX-7062