[FIX] sap.m.Popover: beforeClose now fired correctly on auto-close
If the popover is closing via auto-close mechanism of the popup
we want to fire the beforeClose event even if the focus is still in the
Popup - in case the Popover opener doesn't exist anymore.

There are cases when the popover should be closed from a control (like a
closе button) inside of it, but a moment earlier
its opener gets hidden by other aside logic - this is causing the auto
close to be fired first - before the 'close' caused from
the button press, resulting in that beforeClose event never being fired.

BCP: 2080253679
Change-Id: I0ba27dd34dfda490efc88bf50c6e5d84c93f4d4b