[INTERNAL][FIX] ActionToolbar: Error on destruction
The ManagedObject deletes the information about the content aggregation from
"mAggregations". If "_onContentPropertyChangedOverflowToolbar" is triggered
while the ActionToolbar is being destroyed, getAggregation("content") is called.
The ManagedObject now adds "content: []" to "mAggregations". For all
aggregations that are listed in "mAggregations", the corresponding destructor is
called at a later destruction step. The ActionToolbar throws an error when
"destroyContent" is called.

Change-Id: I1ff1c4b8bc9b318e525e4dc3c7602a3663f04923
BCP: 2080242259