[FIX] sap.m.Select: Fixed out of bounds text in the list items
Problem: Sometimes the text in the second column of the list items is
too long and goes out of bounds of the SelectList (picker). This causes
an unnecessary horizontal scrollbar and other visual issues.

Solution: We force the text in the columns of the list items to be in
bounds of their containers/spans. This is done by CSS styling.

Change-Id: I2f34ea075c840e4999c483818e503a9ede4f4466
BCP: 2080282515
(cherry picked from commit c9bde94bbcd02a5a6ebc70270fbf0a5718fa1fed)
CR-Id: 002075125900003215282020