[INTERNAL] sap.m.Select: Improve accessibility
Problem: Value of selected item was part of Select's label and
it was not distinguishable.

Solution: Refactor the implementation to use native select and set
the value attribute, remove the aria-live property as the native select
handles the announcement of the new value correctly. Set focus and all
aria attributes to the internal select html element.

Change-Id: I463f835f03f17d9514b31a487905dcd471dfc273
BCP: 002075129400003876492020