[INTERNAL] sap.m.Select: Improve accessibility
Problem: Value of selected item was part of Select's label and
it was not distinguishable.

Solution: Refactor the implementation to use native select and set
the value attribute, remove the aria-live property as the native select
handles the announcement of the new value correctly. Set focus and all
aria attributes to the internal select html element.

Important: A platform dependant solution is introduced to announce the new value when
changing with arrows and the dropodown is closed. Only on MacOS and when the select is
focused.

Change-Id: I463f835f03f17d9514b31a487905dcd471dfc273
CR-Id: 002075125900003951852020
BCP: 002075129400003876492020