[FIX] sap.uxap.AnchorBar: CustomAnchorBarButtons no longer hidden
Problem:
For each button of the achorBar menu, the control creates a corresponsing
select-item (to show on phone-view).
However, if the button uses data-binding, the data binding may not have been
resolved yet at the time of creation of the corresponding select-item
=> the final value of the bound property is lost

Solution:
Ensured that binding info is also set to the corresponding select-item.

Note:
For simplicity, we support a setup where the model and the bindingContext
are set globally on the entire objectPage/view, which is the most
common and expected scenario.
Further scenarios (separate model and context set at individual buttons) may be
addressed at a planned refactoring change 4871181)

Change-Id: Id8abd5067e195b5d271035aeb79cdd4e1802c3ba
BCP: 2080270787