[FIX] sap.ui.documentation: resolve bootstrap URL relative to baseURI
When the download functionality rewrites the bootstrap URL of the
index.html of the samples, it should resolve a relative URL relative to
the document.baseURI, not to the window.location. Otherwise, the URL
will be broken when a different <base> URL has been set for the SDK,
e.g. in the SEO variant on openui5 or openui5nightly.

Fixes: https://github.com/SAP/openui5/issues/2964
Change-Id: I18a008739f74465799e1d2f5c396e2f944bab2bb
BCP: 2080270758
BCP: 2080307383