[INTERNAL] sap.m.Input: selectedKey + value- align behavior
We'd like to have a clear understanding on how the
selectedKey + value would behave. Currently, there are lots
of use cases which favor the selectedKey or the value.
However, we've come to a situation where we're not able to support
all the scenarios as some of them are opposite.

Here's the flow that we've aligned on:
- selectedKey: true, value: true => if (matching item) => value = selectedItem.text else value = value

- selectedKey: false, value: true => value

- selectedKey: true, value: false, matchingItem: true => value = selectedItem.text else value = value

JIRA: BGSOFUIRILA-2884

Change-Id: I491fedc1b8a05acc39e0bb3618a378589ec5280a