[INTERNAL][FIX] Tables: Handling of V4 virtual context and busy indication
A sap.m.Table that is bound to a V4 ODataListBinding with "autoExpandSelect"
needs to handle virtual context scenarios to correctly display the NoData
text and the busy indicator.
Downport of I5d795b61c79ce95dc562658581e5bafe407183e0

In a sap.ui.mdc.Table of type "ResponsiveTable", a busy indicator delay of
less than 1000 (default) is required, because updating the items can take a
lot of time (that's a different problem).
If the response is received before the indicator is visible, a long-running
update of items starts without any busy indication. The user has the visual
impression that nothing is happening for a few seconds (in this app) until
suddenly new content appears.
Downport of I90cbedf714583d20f81a2184cb804fa28bc68725

Change-Id: I5d795b61c79ce95dc562658581e5bafe407183e0
CR-Id: 002075125800003252442020
BCP: 2070303752