[FIX] sap.m.Popover, sap.m.MessagePopover: Runtime exceptions cleared
sap.m.Popover:
In some cases the DOM manipulating private function _clearCSSStyle is
being called without having a rendered popup DOM and it was throwing an
exception. Now a new check is added to make sure the Popover is rendered
before executing the rest of the method.

sap.m.MessagePopover:
_setInitialFocus() and _syncMessageView() are sometimes called when no MessageView is
initialized. Additional conditions are added wherever methods of the
MessageView are called to prevent exceptions.

BCP: 2080193375
Change-Id: I8bf1a1455da02ed146f03e67db1333ecea2b8d66