[INTERNAL] FieldHelp: support grouping
For FieldValueHelp grouping could just be defined in the binding of the
table, only some navigation handling needs to be added.

For ListFieldHelp a new ListFieldHelpItem eleent needs to be added to
enable binding of the group name. The grouping is then done in the
binding of the internal list.

Change-Id: I224ae969994a01306bd0ae61e769816721dad5b4
JIRA: CPOUIFDENVER-1550