[FIX] ColumnLayout: ResizeObserver loop limit exceeded error fixed
Use the windows.requestAnimationFrame() API to defer the number of
reflows that are performed per animation frame, to not exceed the
resize observer loop limit.

The resize observer loop limit is usually exceeded when size changes
are made in resize callbacks (implicitly or explicitly) and when the
changes are processed in the same animation frame.

Change-Id: I0871d503861fb5b65c033100beeda50e4d684545
BCP: 2080323528