[INTERNAL] Comply with version 3.0 of the REUSE Specification
See https://reuse.software/spec/
And other requested legal adaptations (cleanup of LICENSE.txt).

Explanation:

LICENSE.txt was cleaned up to only contain the original generic license
text. All third-party components, which were previously mentioned there,
are now listed in THIRDPARTY.txt, along with a copy of their license
texts.
The SAP API clause is now in the .reuse/dep5 file, as well as the also
removed NOTICE.txt. This dep5 file also contains all third-party
information in machine-readable format, as required by REUSE.

Both, THIRDPARTY.txt and .reuse/dep5, are fully generated from the
information maintained in the .library files (generator is right now in
ui5-tools/thirdparty-metadata-generator in the internal GitHub). The
information in the .library files has been extended to encompass
everything needed for generating the two files. In particular, it has
been enriched with the file location patterns, which we did not have
before at all.
The .repo file at root level has the same content as the .library files,
but for resources outside of libraries. While the .library files have
the file patterns in runtime format (they are transformed by the
generator for the dep5 file to be in source format), the paths are
always in source format for the .repo file.

All used licenses were downloaded to the LICENSES folder for the REUSE
tool. The old license copies we had were removed.

As the dep5 file cannot express that single files contain parts with
different copyright, such files are annotated with SPDX
copyright/license information in their headers. The dep5 file does not
contain them, but REUSE will pick up license information from the
headers. The .library files do contain information for such files, but
with a "partial=true" flag. The THIRDPARTY.txt, in contrast to dep5,
does list such partial files as having content from all respective
third-party libraries.

The CLA and some outdated LICENSE copies have been removed and
the README.md adapted (and updated with some newer links).

The Readme.md files of all libraries (mainly needed for the npm release)
have been adapted to mention the THIRDPARTY.txt file.

The guidelines have also been updated to say that modifications to
third-party components should be marked with "BEGIN: MODIFIED BY SAP"
and "END: MODIFIED BY SAP". The older guideline and this one were both
used in parallel (usages of the old one are not cleaned up yet).

The unused AUTHORS file was removed.

The OpenUI5 demokit was adapted to display the new THIRDPARTY.txt in
addition to LICENSE.txt. For the SAPUI5 demokit this is not easily
possible, because third-party modules are added by libraries from others
in upper layers, so it remains loading the information dynamically from
the .library files.

Some download tools in the demokit (for samples) have been adapted to
include the THIRDPARTY.txt file instead of NOTICE.txt and to take them
from the root level, not their outdated local copy, which is now
removed.

The pom.xml was extended to create a THIRDPARTY artifact.

Change-Id: I773d57978dd6beaaaafafd29436e138e73194a65