[INTERNAL] ListBase: Improve rebind handling
On rebind of the items, avoid invalidation before updating the aggregation.

If the items are rebound, they are first destroyed, invalidating the
control. It could now be re-rendered before the data is received and new items
are created. This can happen, for example, when adding columns in the
SmartTable or MDCTable.

Change-Id: Idfd7900ca7b8eb04b65ceb32a662b24b88164178
JIRA: CPOUIFTEAMB-1524