[INTERNAL] sap.f.GridContainer: Fix drag and drop in case of grid invalidation
Solving the following problem:
If the grid is invalidated during drag and drop,
the semantic rendering was breaking.
Also the drag and drop was interrupted.
This also fixes the case when template is used to
create the grid items. Which is the case in the sample
"Drag and Drop From List to GridContainer".

Change-Id: I5551b6f3aa5a235028722f7542f4dbd313ff8b2e
BCP: 2080197575