[INTERNAL]Table: always provide config for inbuilt filtering
The Table will initialize the inner AdaptationFilterBar
in case the first rebind occurs without p13n being active yet
to be able to call 'triggerSearch' on the inner FilterBar.

Once a variant has been created and the app gets restarted,
the AdaptationFilterBar is being created but the 'filterControl'
config for the AdaptationController is only provided once the
personalization has been active for at least once.

To prevent this issue the configuration should always be passed
once the AdaptationFilterBar is being created and not only when
personalization has been used.

Change-Id: Ie8e8ca1443ff5387b2a49c4f5dac83eb6993fdd5