[INTERNAL] AggregationBaseDelegate: harmonize flex specific hooks
The MDC Controls Table, Chart and FilterBar behave similar
when it comes to the handling of the inner default aggregations for:

- Table: columns
- Chart: items
- FilterBar: filterItems

These list like aggregations behave similar in terms of personalization
and lifecycle handling via add/remove hooks which are called out
of the implementing delegate.

The prior methods beforeAddColumnFlex, beforeAddFilterFlex, beforeAddItem
and afterRemoveColumnFlex, afterRemoveColumnFlex and afterRemoveItem
have a similar signature and similar semantic.

To align these three control delegates a derivation "AggregationBaseDelegate"
to the "BaseDelegate" is being used.

Change-Id: I62552ef1b18f87d17b22dfc5a7c0efc373f341da