[FIX] sap.m.RadioButton: Removed custom setter for selected property
Fixing the issue where RadioButton of one groupName, with selected=true,
gets unchecked by another selected=true RadioButton from a different groupName.

The code that maintains a RadioButton to be only in one groupName at a time
has been moved to occur in the "BeforeRendering" phase, where the groupName
is set and won't change before rendering the RadioButton.

Change-Id: I4a62c08fb86749e14433ac5063c6a45394de50dc
BCP: 2080296975