[INTERNAL] sap.ui.integration: Consider deeply nested wrappers during ready check
When nested editors register at their parent editor, a check is
executed to determine if the element is an actual wrapper or the
template of an aggregation binding.
Directly nested editors which are not part of a bound aggregation
might miss a parent somewhere up their hierarchy because the fragment
they are contained in wasn't added to any aggregation of the
complex editor yet.
This might confuse the check for template elements, thus parent
changes of the nested editor have to be observed to re-execute the
check once a parent was added.
So far we only considered direct parent changes of the wrappers,
with this change we also check parent changes up in the hierarchy to
cover cases where the wrappers are deeply nested inside the fragment.

Change-Id: I353b0e6c7bba1b7a5e0c30c0f7e0bcdf1fa155b5
BCP: 2080336088