[INTERNAL] GrowingEnablement: fix applyChunk for Table with hidden columns
When the Table is bound at least 1 column is visible, the
items in the table are created and rendered as expected. Later, all the columns
are set to visible=false, and the user creates data for the table, in this
case, the GrowingEnablement would create empty rows which lead to unexpected
behavior. So to fix the issue, the GrowingEnablement, now checks whether at least
1 column is visible in the table and then performs DOM update in the table, else
the changes are collected in chunks and applied to the table once a column becomes
visible and a rebind is performed.
BCP: 2080250160

Change-Id: I912616b3e69f1cf3a2504a3846872e9d0e46cdac