[INTERNAL][FIX] sap.m.OverflowToolbar: Corrected measurements changes with flexible content
This reverts commit 1218371bd3196c64fbd6a97584fb993797cc4ab6.

Reason for revert: There is a side effect, when resizing the Browser window - so that the GenericTags go in the Popover of the OverlowToolbar, and after that - resize the Browser to bigger size again. In this case, when Browser is shrinked, less space is allocated in the DynamicPageTitle for the OverflowToolbar with GenericTags (as with this change they get cached again). If we remove this logic, the old cached value (when the GenericTag has enough space), won't be changed, and when the GenericTag gets out of the Popover, the allocated space for it would be the same as before (so no cutting and no shrinking).

Change-Id: I727d6a6243cb5e45278cd64b552373ac412503a0
BCP: 2080205481