[INTERNAL] sap.uxap.ObjectPageSection: Update invisible text when new title is set
Background: When there is only one subsection and the section does not have title,
the subsection's title is set to the _internalTitle property and then is used for
aria-labelledby attribute when the whole section takes the focus.

Problem: When new title is set to the subsection, in the onBeforeRendering hook of
ObjectPageSection there is a check to not update the ariaLabelledBy association
if it's already set. Therefore the value of the InvisibleText holding the aria-labelledby
is not updated with new title.

Solution: Update the InvisibleText 'text' property with the new title value in onBeforeRendering hook.

Change-Id: I7b425384463b95117f106e6e2d979e660c7a4728
BCP: 2080231822