[INTERNAL] sap.m.Select: Remove aria-hidden from select and hide options
Problem: Select's options are visible to the accessibility tree but,
not really displayed to the user, so validators complain about their styles.
Also 'aria-disabled' is not allowed on select element, but only 'disabled'.

Solution: Hide the options with the 'hidden' attribute, remove 'aria-disabled'
from select and leave only 'disabled'.

Change-Id: I724321015cd8a4421516e422c104c29e33c12cdb
BCP: 002075129400003876492020