[INTERNAL] StashedControlSupport: Rework of existing private API
The existing private stashing API has been reworked.
Previously, during XML processing, a StashedControl was created as
a separate instance which was not inserted into the control tree.

With this reworked version, a stashed control is now represented
by an empty placeholder control. 'Empty' meaning no child-nodes or
aggregation content. PropertyBindings are still available on the
placeholder controls, e.g. for titles or i18n texts.
The placeholder controls are of the same class as the final control.

On unstashing the placeholder control is replaced with the final
control instance. The final control is placed inside the
parent aggregation of the placeholder at the placeholder's index.
The placeholder is then destroyed.

JIRA: CPOUI5FRAMEWORK-131
Change-Id: I21cb78dddc7c84549c89035a4b12057adc78053d