[FIX] sap.f.GridContainer: Click on a button inside is now working
We are fixing the following issue:
When any clickable content inside an item wrapper (for example button)
was clicked, the ItemNavigation would move the focus from the button
to the wrapper. If the wrapper was not fully visible, the view would
be scrolled to show the wrapper, resulting in prevented action.

Change-Id: I072cf69bd27495a02da208a617ec903aeb78bbc2
BCP: 2080256996