[FIX] sap/ui/thirdparty/datajs: make OData the default AMD export
When the UI5 loader is used, its shims ensure that the global 'OData' is
used as export of the 'sap/ui/thirdparty/datajs' module.

When UI5 exposes itself as standard AMD loader, datajs will export two
modules: 'datajs' and 'OData'. As none of them matches the requested
name 'sap/ui/thirdparty/datajs', the loader will pick the first export
and export it a 2nd time, using the requested name as an alias.

For datajs, this led to the wrong export (datajs instead of OData). As a
fix, the order could be reversed.

To make the solution more obvious, an explicit export to
'sap/ui/thirdparty/datajs' has beeen added.

Change-Id: I76e117b2e902e58e8a1fbf4fdb74b2cd49197e87
Fixes: https://github.com/SAP/openui5/issues/2957
BCP: 2080356740