[INTERNAL] sap.m.Table: Correct firing of popinChanged event
According to the BCP incident we've realized that the event does not get
fired when filtering the table that comes up with an empty result. For
the sap.m.Table as stand-alone this is not really necessary but it
becomes
as soon as the sap.m.Table is part of the SmartTable or mdcTable and the
property showDetailsButton is set to 'true'.
Improving stability of QUnit test in mdcTable w.r.t popinChanged event and
showDetails button visibility.
BCP: 2080218151

Change-Id: Ife474c2953771b30a648e9afd871b357e8cbd743