[FIX] sap.m.MultiInput: DOM is now updated after tokens are destroyed
After the sap.m.Tokenizer refactoring that was part of our Inputs
alignment process the Tokenizer no longer invalidates the MultiInput as
it was an anti-pattern (child controlling its parent).

However because of that when the tokens aggregation is destroyed
the MultiInput was no longer getting rerendered and it's DOM was not
updated, more specifically the 'sapMMultiInputHasTokens' class was not
removed.

ManagedObjectObserver is introduced to observe changes in the forwarded
to the Tokenizer tokens aggregation, because no events are fired, which
is correct since the firing of events must be only in a result of a user
interaction, which we do not necessarily have in this case.

When the tokens get destroyed the observer invalidates the MultiInput
so that the correct class will be applied/removed from it.

BCP: 2080358002
Change-Id: Idcd528b5d581d9958bff63277d4110c9020154ce