[FIX] sap.m.Panel: Content is now always the same height as the Panel
Previously the panel's content height was taken from the panel itself
causing issues in cases when the panel is set to a fixed percentage
height. For example - if the panel is set to 50% height its content
was also set to 50% height, making the content 25% of the height of
the panel's parent container.

BCP: 2080360221
Change-Id: Idcd528b5d581d9958bff63277d4110c9020152ff