[INTERNAL][FIX] Table: Right border of the last visible fixed column
Another fix was provided recently for the case when the last fixed column is invisible. In this case
another fixed column is set to invisible and the border was not shown.

The logic behind the fixedColumnCount property was changed. The value of the property is set to the
value assigned via the setFixedColumnCount method. The renderer uses a computed value of the
fixedColumnCount based on the property value itself and the header spans.

Change-Id: Ic38b5127f832454733273cd9718eb25f926b26d2
BCP: 002075129400005293202020