[INTERNAL] sap.ui.codeeditor: library.js should use sap.ui.getCore()
In a sync bootstrap scenario or when the 'xx-bootTask' feature is used,
sap/ui/core/Core.js and the **/library.js modules of bootstrap libraries
can form a cyclic dependency. Therefore, the library.js receives 
undefined as export of sap/ui/core/Core.js.

Calling initLibrary() on that export therefore might fail.

library.js modules or other modules that are likely to be required in
an xx-bootTask and all their transitive dependencies therefore
should call sap.ui.getCore() instead of using the export, for better
robustness.

Change-Id: Icba3d38971d56c55cb5f72a6fd580e05d7842867