[INTERNAL]: Field: don't update output if update in progress
If a Field value was changed the update of the internal condition is
triggered asynchronously. (To update value and additionalValue togeter.)
If in the meantime the assigned FieldHelp gets a data update and fires
it's dataUpdate event a new formatting of the output is triggered. If at
this point of time the condition of the Field waits to be updated a
formatting of the old value is triggered. That's not needed and might
lead to unneeded requests or FormatExceptions.


Change-Id: I78eb935748a104a50c0582316a97680bc7d99193