[INTERNAL][FIX] sap.m.Select: Handle focus restore to select and prevent native dropdown
Problem 1: When focus is about to leave the select (a SelectList item is
clicked) we restore the focus to the hidden select. While this works for
the first time, after that the onsapfocusleave handler is no longer called. In
some browsers this actually leads to infinite recursive call,
because the onsapfocusleave handler is called every time the
this._getHiddenSelect().focus()

Solution: Restoring the focus by calling the focus() method on the
Select control itself does work and the onsapfocusleave handler is again called
if a SelectList item is clicked.

Problem 2: Native select has a sapUiPseudoInvisibleText class
which makes it invisible to a mouse click, but when an mousedown actions
is fired on it the native dropdown listbox is opened.

Solution: Prevent the default behavior in a mousedown hook handler.

Change-Id: I88d958dd4d88c071a47d5b256b449e245d2b8d8b
BCP: 2080366605