[FIX] sap.f.DynamicPage: Fixed incorrect display of pin button
Problem: The pin button is displayed when the height of the DynamicPage
is changed and 'preserveHeaderStateOnScroll' property is 'true'.

Root Cause: Whether or not to display the pin button, we depend on a
boolean returned by the '_preserveHeaderStateOnScroll' private function.
The returned value is sometimes incorrect when the height of the
DynamicPage is changed.

Solution: We make sure to update the '_preserveHeaderStateOnScroll'
boolean return value status, when the height of the DynamicPage is
changed.

Change-Id: I1da671ab52474b115d4e2a5ba3e27640ba99cfb4
BCP: 2080370980