[INTERNAL] sap.m.Select: Change role to listbox
Problem: After removing the aria-live attribute,
(I463f835f03f17d9514b31a487905dcd471dfc273)
the new selection was not read out with some
screen readers (VoiceOver) on Safari when dropdown is open.

Solution: Changing the role to listbox
the new value is read out which is also
according to the specification.

Change-Id: I83116939c3e7811b431ca5fd16594dbab4e59485
BCP: N/A