[INTERNAL][FIX] sap.ui.fl maps the version number correct
With the previous refactoring to align the naming of the
sap.ui.fl.Versions from .versionName to .version, also the expected
parameter from a back end response was changed.
Since the back end still sends the property versionNumber, the
functionality is broken.
This commit maps the response correctly (sending the versionNumber) to
the client parameter (everywhere version).

Change-Id: Id036862820c56bae3ab0da852b20327d154335e0
BCP: N/A
CR-Id: 002075125800003085062020