[INTERNAL] Table: Make use of the lodash "throttle" function in table utilities
Use the "throttle" function from lodash inside TableUtils.throttle, adding the
option for asynchronous leading invocations.
Replace the no longer required TableUtils.debounce function with a function to
throttle frame-wise (requestAnimationFrame), which was the only used feature of
this utility. For other cases, if required, lodash "debounce" can be used.

Change-Id: Ic911b5b9dd33d567a83add0e2f2b236496a131ad