[FIX] mdc.FieldValueHelp: Value not displayed in edit mode fixed
The issue:
The value of the text input field is not always displayed in edit
mode when `in` parameters are used.

The root cause of the issue:
In the _getParameterBinding() function of the FieldValueHelp.js
module, the sap.ui.model.odata.v4.Context#getProperty() method
is invoked after the binding context is destroyed, causing an
exception to be raised.

Call stack of the handled exception in the
ManagedObject.prototype#updateProperty method:

TypeError: Cannot read property 'checkSuspended' of undefined
    at constructor.Context.getProperty (Context-dbg.js?eval:623)
    at fnClass._getParameterBinding (FieldValueHelp-dbg.js?eval:920)
    at fnClass.FieldValueHelp._getTextOrKey (FieldValueHelp-dbg.js?eval:882)
    at fnClass.eval (FieldHelpBase-dbg.js?eval:1267)
    at eval (SyncPromise-dbg.js?eval:309)
    at call (SyncPromise-dbg.js?eval:60)
    at new SyncPromise (SyncPromise-dbg.js?eval:225)
    at SyncPromise.then (SyncPromise-dbg.js?eval:308)
    at fnClass._getTextAndKeyAfterContentLoaded (FieldHelpBase-dbg.js?eval:1266)
    at fnClass._handleContentLoadedForTextAndKey (FieldHelpBase-dbg.js?eval:1245)

The solution:
The control flow in the _getParameterBinding() function is shifted
to the last `else` statement to await until the binding of the `in`
parameter is propagated.

Change-Id: Ifee3a2f795d0b1d59ec055a57d9073709dc0a62b
BCP: 2070284113