[INTERNAL][FIX] sap.m.Select: Hide arrow and label from accessibility tree
Problem: When select is operated with VoiceOver under mobile phone
the label text and arrow icon are selectable by tapping on them.
This leads to the actual select being very hard to be select, therefore
the aria attributes and selected value are not read out.

Solution: Set aria-hidden to the arrow and label divs to hide them from
accessibility tree.

Change-Id: I430e401840a294c8bcce282154b6597e06052e61
BCP: 002075129500005360892020