[FIX] sap.m.table: Fixed failing assertion for the renderer
Some lines in the renderer was causing assertion failures from the renderer API. The class
method of the renderer is supposed to have only one class but in the previously there
was an option to send more than one class as a string.

We split the string to make sure that even if there is more than one class, the
renderer will pass one by one.

BCP: 2080371005
Change-Id: I941ec7a954c774a8d09073db93b938fafd27513f