[FIX] sap.f.DynamicPage: Issue with wrong scroll position is fixed
Problem: When the header is snapped with scroll, the scroll position is
getting additionally adjusted (on height change) in a way that results in
a wrong scroll position and in some cases with sap.m.Table with
growingScrollToLoad enabled it leads to loading of more items.

Solution: Additionally adjusting the scroll position when there is change
in header's height, only when it is visible.

Change-Id: Icc8d172079e4b90687a768556c88e0d2291783a0
BCP: 2070010871
CR-Id: 002075125800003294312020
(cherry picked from commit 14f1014a903b4de1306fe4108bbf44e108329660)