[FIX] sap.f.DynamicPage: Issue with wrong scroll position is fixed
Problem: When the header is snapped with scroll, the scroll position is
getting additionally adjusted (on height change) in a way that results in
a wrong scroll position.
In some cases with sap.m.Table with growingScrollToLoad enabled, this leads
to loading of more items of the  Table (when not needed). As a result, a
second roundtrip for requesting data is send, which leads to performance
violation. Performance problem is already observed in multiple applications.

Solution: The scroll position is additionally adjusted upon change in header's
height only when it is visible. This way, the page is not scrolled, when not
needed, and there is no second data request.

Change-Id: Icc8d172079e4b90687a768556c88e0d2291783a0
BCP: 2070010871
CR-Id: 002075125800003294362020
(cherry picked from commit 14f1014a903b4de1306fe4108bbf44e108329660)