[INTERNAL] MDC Table: Column Header Labelling Association
The MDC Column creates DOM element in the static area which has the
columnHeader label as the text of the DOM element and ID same as that
of the MDC Column.

In complex scenarios, the application has to associate
with the MDC Column ID  which internally associates with the
DOM element that makes it appear like it is associated with the MDC
columnHeader Label.

JIRA: CPOUIFTEAMB-1493
Change-Id: I0b6e209cea02c03b2811c2524b4f792ec41618f0