[FIX] sap.ui.model.analytics.AnalyticalTreeBindingAdapter#setNumberOfExpandedLevels: expand only as many levels as grouped dimensions are available
A dimension (e.g. CompanyCode) and its text (e.g. CompanyCodeName) can
be selected as grouped columns. For the analytical table each grouped
column is counted as a group level, but for the AnalyticalBinding only
the dimensions are counted as groups. CompanyCode and CompanyCodeName
represent the same dimension so the AnalyticalBindings counts these two
grouped columns only as one group.
AnalyticalTreeBindingAdapter#setNumberOfExpandedLevels needs to restrict
the maximum number of expand levels to load the leaf entries properly.

http://localhost:8080/testsuite/test-resources/sap/ui/core/qunit/analytics/testsuite4analytics.qunit.html

PS1: Done
PS2-3: Review comments

Change-Id: I756d24c2a9a7e62d9a890456bdefd311d6b705ce
BCP: 002075129500005340402020