[FIX] format/NumberFormat: consider control characters in minus sign
For Right-To-Left locales it is possible that the minus-sign may contain
additional control characters. For currency patterns which have a second
part for negative numbers, the first character was stripped, which can
be the control character (e.g. for HE locale). This could lead to
doubled minus signs. To avoid this, the length of the minus sign is now
considered.

BCP: 002075129400005462222020
Change-Id: I0718ff54943c85a8168ede6ebc846b6599c19e4b