[FIX] sap.m.OverflowToolbar: Unnecessary caching of width is prevented
Problem: GenericTags are getting truncated, when placed in
OverflowToolbar, part of the "content" aggregation of DynamicPageTitle,
when screen size is made smaller.

Solution: Problem comes from a recent change in the OverflowToolbar,
related to new caching of controls with relative widths (specified in
%). Additional check is implemented, so that now controls without
specified with ("" is returned as width value), do not require new width
caching.

Change-Id: I6bdfb04115612cca3fc5b4a38751a44b7585496f
BCP: 2070295397