[FIX] sap.m.Select: Fixed missing focus on open
Problem: The browser click event triggers focus on the Select control after the popup caches
the previously focused element. That way, after the popup is closed it returns focus
on the previously focused element,which is not the Select control as it should be.

Solution: We trigger focus on opening of Select before calling the "open" method of the
popup to make sure it is focused.

Change-Id: Ib2dbc2e28a6c088bbf041c6f76f6278d515ece23