[INTERNAL] m.Column: changing visible should not affect visibility via media queries
If a column is hidden via minScreenWidth and the visible property is toggled,
then the setDisplay function in the Column made the column visible without
respecting the media info. This is now fixed, by adding a special check in the
setDisplay function for isHidden function which determines the visibilty of a
column based on its media information.
BCP: 2070300045

Change-Id: Ifa5a853062a19e2c001e4fff406d303e627e598f