[FIX] mdc.Chart: Breadcrumbs are not displayed on back navigation fixed
The issue:
- When the Chart control is initially rendered, the breadcrumbs are
displayed as expected. But after navigating to a different screen,
and then navigating back to the initial screen, the breadcrumbs
are no longer displayed.

The root cause:
In some scenarios, the sap.ui.mdc.Chart.prototype#_createDrillBreadcrumbs
method was called before the Chart hidden aggregation was created,
causing the following .setAggregation() call in the DrillStackHandler
module to be bypassed:

[...]
var oInnerChart = oChart.getAggregation("_chart");

if (oInnerChart) {
	var oDrillBreadcrumbs = new Breadcrumbs();
	oChart.setAggregation("_breadcrumbs", oDrillBreadcrumbs);
[...]

The solution:
Create the breadcrumbs after the Chart aggregation is created.

Unrelated issue fixed:
Invalid selection mode values are not propagated to the inner
Chart control when the sap.ui.mdc.Chart#setSelectionMode()
method is called.

Change-Id: Ica0dd0b7193c82f26461393e260cc7cf6e7fb674
BCP: 2080374199